Generalize the info methods
These methods are broadly the same, the only difference being the model
object that is being displayed
