+++ lib/cloudware/commands/infos/machine.rb
@@ -5,27 +5,9 @@ require 'tty-table'
 module Cloudware
   module Commands
     module Infos
-      class Machine < Command
-        include Concerns::ExistingDeployment
-        include Concerns::Table
-        attr_reader :machine_name, :deployment_name
-
-        def run
-          @deployment_name = options.deployment
-          @machine_name = argv[0]
-          machine.tags.each { |row| table << row }
-          puts render_table
-        end
-
-        private
-
-        def machine
-          Models::Machine.new(name: machine_name, deployment: deployment)
-        end
-        memoize :machine
-
-        def table_header
-          ['Tag', 'Value']
+      class Machine < Info
+        def model_class
+          Models::Machine
         end
       end
     end
